fix db models
